// src/app/models/user.ts
export class User {
  UserGUID: string;
  UserName: string;
  FirstName: string;
  LastName: string;
  RoleMetaNameList: string;
  AccountNumber: string;
  AccountName: string;
  AccountEmail: string;
  DeliveryAddress: string;
  DeliveryCity: string;
  DeliveryPostalCode: string;
  UserEmailAddress: string;
  UserPhone1: string;
  UserPhone2: string;
  BaseCategory: string;
  AddressLine1: string;
  AddressLine2: string;
  UserExternalID: string;
  UserID: number;
  UserGroup: string;
  UserId: string;
  UserType: string;
  InvoiceToCustomerID: string;
  DeliverToCustomerID: string;
  AdminPersonalID: string;
}

export class RoleMetaName {
  Id: number;
  Name: string;
}

export class RoleMetaNames {
  Id: number;
  Name: string;
}